pyproject.toml
[project]
name = "pygame_intro"
version = "1.3.1"
authors = [{ name = "AntonisDevStuff", email = "contact@antonisdevstuff.com" }]
description = "A intro module for pygame Community Edition."
keywords = ["pygame", "intro", "splash screen"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.0",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: pygame",
]
[project.urls]
Repository = "https://github.com/antonisdevstuff/pygame-intro"
Issues = "https://github.com/antonisdevstuff/pygame-intro/issues"
Releases = "https://github.com/antonisdevstuff/pygame-intro/releases"

[tool.setuptools]
package-dir = { "pygame_intro" = "src" }
packages = ["pygame_intro"]

[tool.setuptools.package-data]
pygame_intro = ["data/**"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
